[PLATFORM-303] Refactor exporter
