+++ .gitignore
@@ -1,4 +1,5 @@
-tests/playbok.retry
+tests/playbook.retry
+tests/.cache
 tests/__pycache__
 .molecule
 .vagrant
