+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a changelog](https://github.com/olivierlacan/keep-a-changelog).
 
 
-### Unreleased
-
-## 1.1.0
+## [Unreleased](https://github.com/idealista-tech/prometheus_jmx_exporter-role/tree/develop)
 ### Added
-- [PLATFORM-162](http://jira.sys.idealista/browse/PLATFORM-162) Utilizar systemd con ansible, reducir el número de métricas que generamos para java y tomcat
+- *[PLATFORM-303](http://jira.sys.idealista/browse/PLATFORM-303) Exporter refactor* @jmonterrubio
+
+## [1.1.0]
+### Added
+- *[PLATFORM-162](http://jira.sys.idealista/browse/PLATFORM-162) Utilizar systemd con ansible, reducir el número de métricas que generamos para java y tomcat*
