+++ handlers/main.yml
@@ -1,6 +1,7 @@
 ---
+
 - name: restart jmx_exporter
   systemd:
     name: jmx_exporter
     state: restarted
-  when: jmx_exporter_enabled == 'yes'
+  when: jmx_exporter_service_state != 'stopped'
