+++ meta/main.yml
@@ -1,4 +1,16 @@
 ---
+
 galaxy_info:
-  author: Alberto Garcia
-  description: Prometheus jmx exporter role
+  company: Idealista S.A.
+  description: Prometheus jmx exporter
+  min_ansible_version: 2.2
+  license: Apache 2.0
+  platforms:
+    - name: Debian
+      versions:
+        - jessie
+  categories:
+    - monitoring
+
+dependencies:
+  - java
