+++ molecule.yml
@@ -21,13 +21,21 @@ molecule:
   # directory containing group_vars to use with ansible
   group_vars: ../tests/group_vars
 
+dependency:
+  name: galaxy
+  requirements_file: tests/requirements.yml
+
 # ansible related configuration
 ansible:
   playbook: tests/playbook.yml
 
 # configuration options for the internal call to vagrant
 vagrant:
-  # molecule's --platform option will look for these names
+  # molecules --platform option will look for these names
+  raw_config_args:
+  - "landrush.enabled = true"
+  - "landrush.tld = ''"
+  - "landrush.guest_redirect_dns = false"
   platforms:
     - name: Debian
       box: debian/jessie64
@@ -40,12 +48,7 @@ vagrant:
         cpus: 1
 
   instances:
-    - name: jmxExporterDown
-      interfaces:
-        - network_name: private_network
-          type: dhcp
-          auto_config: true
-    - name: jmxExporterUp
+    - name: jmxExporter
       ansible_groups:
         - group01
       interfaces:
