+++ tasks/service.yml
@@ -1,0 +1,8 @@
+---
+
+- name: JMX_EXPORTER | Configuring service
+  systemd:
+    name: jmx_exporter
+    state: "{{ jmx_exporter_service_state }}"
+    enabled: "{{ jmx_exporter_service_enabled }}"
+    daemon_reload: yes
