+++ templates/jmx_exporter.service.j2
@@ -4,11 +4,20 @@ Requires=network-online.target
 After=network-online.target
 
 [Service]
-Restart=on-failure
-ExecStart=/bin/sh -c "jmx_exporter {{ jmx_exporter_port }} {{ jmx_exporter_config_file }}"
+PrivateTmp=yes
 User={{ jmx_exporter_user }}
+Group={{ jmx_exporter_group }}
+RuntimeDirectory=jmx_exporter
+Environment='JAVA_OPTS=-Dcom.sun.management.jmxremote.port={{ jmx_exporter_internal_port }} \
+    -Dcom.sun.management.jmxremote.authenticate=false \
+    -Dcom.sun.management.jmxremote.ssl=false'
+ExecStart=/bin/sh -c "/usr/bin/java $JAVA_OPTS -jar {{ jmx_exporter_root_path }}/jmx_exporter.jar {{ jmx_exporter_port }} {{ jmx_exporter_conf_path }}/jmx_config.yml > {{ jmx_exporter_log_path }}/{{ jmx_exporter_log_file }} 2>&1"
+ExecStartPost=/bin/sh -c "echo $MAINPID > /run/jmx_exporter/jmx_exporter.pid"
 ExecReload=/bin/kill -HUP $MAINPID
+WorkingDirectory={{ jmx_exporter_root_path }}
+PIDFile=/run/jmx_exporter/jmx_exporter.pid
 KillSignal=SIGTERM
+SuccessExitStatus=143
 
 [Install]
 WantedBy=multi-user.target
