--- tests/playbook.retry
@@ -1,2 +1,0 @@
-jmxExporterDown
-jmxExporterUp
