+++ tests/requirements.yml
@@ -1,0 +1,4 @@
+- src: https://github.com/idealista-tech/java-role
+  scm: git
+  version: 1.0.0
+  name: java
