+++ vars/main.yml
@@ -1,0 +1,6 @@
+---
+
+jmx_exporter_required_libs: []
+
+jmx_exporter_bin_name: "jmx_prometheus_httpserver-{{ jmx_exporter_version }}-jar-with-dependencies.jar"
+jmx_exporter_url: "https://repo1.maven.org/maven2/io/prometheus/jmx/jmx_prometheus_httpserver/{{ jmx_exporter_version }}/{{ jmx_exporter_bin_name }}"
